gcp_cloud_functions/requirements.txt
functions-framework==3.*
google-auth==2.6.2
requests==2.27.1
requests-toolbelt==0.10.0